Move hook to the separated file
